Changes to implement background color at the paragraph level in RTF output.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@196608 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/rtf/rtflib/rtfdoc/RtfColorTable.java
@@ -168,9 +168,9 @@ public class RtfColorTable {
          * @param name a named color
          * @return the RTF number of a named color, or null if name not found
          */
-    public Integer getColorNumber (String name) {
-            return (Integer)namedColors.get(name.toLowerCase());
-        }
+    public int getColorNumber (String name) {
+        return ((Integer)namedColors.get(name.toLowerCase())).intValue();
+    }
 
     /**
      * Gets the number of color in the color table
